refactor: 내부 api 분리
